Patch DumpNativeStack() to include GNU Build IDs
 #### Issues Addressed

This patch implements an alternative DumpNativeStack() implementation.

In contrast to the libart's original code, this will also dump the GNU Build IDs for each frame,
which is necessary to be able to desymbolicate the frames accurately.

The implementation is based on code from tombstone.cpp and BacktraceCurrent.cpp
from the platform/system/core repo.

 #### Summary of Changes

- Exclude existing implementation using `#if 0`
- Add alternative DumpNativeStack()

 #### Test Plan

- Build Android for Cuttlefish
- Trigger an ANR using Android test app
- Run `adb bugreport`
- Look at the ANR report and observe the backtraces look good and include GNU Build IDs for each frame

Note: I didn't see any automated tests in the art repo that look like they exercised the dumping